Ultra Bot created!
Ultra Bot is a knowledge based bot which plays the highest card with 100% confidence level. This is achieved using logic based statements. As of now, Ultra Bot is atleast as strong as RDeep. Do note that Ultra Bot has some logical statements as well, but these statements would be further minimized after Monday.

Also, human does not work as of now!